LayoutTests/js/regexp-overflow-too-big-expected.txt

Tests a regular expression that hits the recursion limit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new RegExp(complexPattern + complexPattern).exec(complexInput + complexInput) is null
PASS successfullyParsed is true

TEST COMPLETE
